@@ -12,43 +12,43 @@
     "@vueuse/core": "^10.7.2",
     "lodash-es": "^4.17.21",
     "pinia": "^2.1.7",
-    "vue": "^3.3.4",
+    "vue": "^3.4.15",
     "vue3-text-clamp": "^0.1.2"
   },
   "devDependencies": {
     "@lehoczky/prettier-config": "^1.1.1",
-    "@nx/eslint": "17.3.1",
-    "@nx/eslint-plugin": "17.3.1",
-    "@nx/js": "17.3.1",
-    "@nx/vite": "17.3.1",
-    "@nx/vue": "17.3.1",
-    "@nx/web": "17.3.1",
-    "@nx/workspace": "17.3.1",
-    "@swc-node/register": "~1.6.7",
-    "@swc/core": "~1.3.85",
-    "@swc/helpers": "~0.5.2",
+    "@nx/eslint": "18.0.0",
+    "@nx/eslint-plugin": "18.0.0",
+    "@nx/js": "^18.0.0",
+    "@nx/vite": "^18.0.0",
+    "@nx/vue": "^18.0.0",
+    "@nx/web": "^18.0.0",
+    "@nx/workspace": "^18.0.0",
+    "@swc-node/register": "~1.6.8",
+    "@swc/core": "~1.3.107",
+    "@swc/helpers": "~0.5.3",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "18.16.9",
-    "@typescript-eslint/eslint-plugin": "^6.13.2",
-    "@typescript-eslint/parser": "^6.13.2",
-    "@vitejs/plugin-vue": "^4.5.0",
-    "@vue/eslint-config-prettier": "7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "autoprefixer": "10.4.13",
+    "@types/node": "^18.19.14",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
+    "@vitejs/plugin-vue": "^5.0.3",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "autoprefixer": "^10.4.17",
     "consola": "^3.2.3",
-    "eslint": "~8.48.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-vue": "^9.16.1",
-    "nx": "17.3.1",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-vue": "^9.21.1",
+    "nx": "^18.0.0",
     "ora": "^8.0.1",
     "playwright": "^1.41.2",
-    "postcss": "8.4.21",
+    "postcss": "^8.4.33",
     "prettier": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.11",
-    "tailwindcss": "3.2.7",
-    "typescript": "~5.3.2",
+    "tailwindcss": "^3.4.1",
+    "typescript": "~5.3.3",
     "unplugin-auto-import": "^0.17.5",
-    "vite": "^5.0.0",
-    "vue-tsc": "^1.8.8"
+    "vite": "^5.0.12",
+    "vue-tsc": "^1.8.27"
   }
 }
